requirements.txt
Django==1.6.2
Jinja2==2.7.3
Markdown==2.4
MarkupSafe==0.23
Pygments==1.6
South==0.8.4
Sphinx==1.2.2
django-debug-toolbar==1.2.1
django-localflavor==1.0
django-mailgun==0.2.2
django-model-utils==2.1.1
django-widget-tweaks==1.3
docutils==0.11
mercurial==2.9
mock==1.0.1
model-mommy==1.2.1
psycopg2==2.5.2
requests==2.2.1
six==1.3.0
smartypants==1.8.3
sqlparse==0.1.11
stevedore==0.14.1
typogrify==2.0.4
virtualenv==1.11.4
virtualenv-clone==0.2.4
virtualenvwrapper==4.2
wsgiref==0.1.2
